"Add create and store methods for AuteurController and related views"
This diff adds the ability to create and store new authors in the application. It includes the following changes:

* Adds the `create` and `store` methods to the `AuteurController`.
* Creates a new view for creating authors at `resources/views/Biblio/createAuteur.blade.php`.
* Adds a link to create a new author on the index page at `resources/views/Biblio/index.blade.php`.
* Adds a new route for creating and storing authors in `routes/web.php`.

The changes also include some updates to the existing `LivreController` and `routes/web.php` files to move the "Create a new book" link into a separate `div` element, allowing for easier styling and layout.