"Add 'remember me' functionality to user login and migration; update UserController and create users table migration"
This commit adds the ability for users to be remembered when logging in, and updates the associated code in the UserController and the migration for the users table. The 'remember\_token' column has been added to the users table migration, and the Auth::attempt method in the UserController has been updated to accept a second argument for the 'remember me' functionality.